+++ seq2mid/src/seqq2mid/seqq2mid.c
@@ -47,7 +47,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 		fprintf(stderr, "Error: Invalid signature\n");
 		return false;
 	}
-	seqStartOffset = ftell(fSEQq);
+	seqStartOffset = (int) ftell(fSEQq);
 	fseek(fSEQq, 4, SEEK_CUR);
 
 	// read qQES header
@@ -127,13 +127,13 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 	// event stream
 	while (1)
 	{
-		seqEventAddr = ftell(fSEQq);
+		seqEventAddr = (int) ftell(fSEQq);
 
 		// delta-time
 		seqDelta = fgetvb(fSEQq);
 		if (seqDelta == EOF)
 		{
-			fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", ftell(fSEQq));
+			fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", (int) ftell(fSEQq));
 			break;
 		}
 
@@ -141,7 +141,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 		seqByte = fgetc(fSEQq);
 		if (seqByte == EOF)
 		{
-			fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", ftell(fSEQq));
+			fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", (int) ftell(fSEQq));
 			break;
 		}
 		// check the end marker!
@@ -170,7 +170,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 
 			if (seqOpcode < 0x80)
 			{
-				fprintf(stderr, "Error: Unexpected opcode at 0x%08X\n", ftell(fSEQq));
+				fprintf(stderr, "Error: Unexpected opcode at 0x%08X\n", (int) ftell(fSEQq));
 				break;
 			}
 
@@ -208,7 +208,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 							{
 								printf("\n");
 							}
-							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", ftell(fSEQq));
+							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", (int) ftell(fSEQq));
 							goto quit_main_loop;
 						}
 						fputvb(seqSysExLength, fMIDI);
@@ -226,16 +226,16 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 						{
 							if (seqSysExLength > sizeof(s))
 							{
-								fprintf(stderr, "Error: Too long message at 0x%08X (the message may be valid, but this tool does not support it because of laziness.)\n", ftell(fSEQq));
+								fprintf(stderr, "Error: Too long message at 0x%08X (the message may be valid, but this tool does not support it because of laziness.)\n", (int) ftell(fSEQq));
 								goto quit_main_loop;
 							}
 
 							if (fread(s, seqSysExLength, 1, fSEQq) != 1) {
-								fprintf(stderr, "Error: File read error at 0x%08X\n", ftell(fSEQq));
+								fprintf(stderr, "Error: File read error at 0x%08X\n", (int) ftell(fSEQq));
 								goto quit_main_loop;
 							}
 							if (fwrite(s, seqSysExLength, 1, fMIDI) != 1) {
-								fprintf(stderr, "Error: File write error at 0x%08X\n", ftell(fMIDI));
+								fprintf(stderr, "Error: File write error at 0x%08X\n", (int) ftell(fMIDI));
 								goto quit_main_loop;
 							}
 							if (!quiet)
@@ -269,7 +269,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 						seqMetaType = fgetc(fSEQq);
 						if (seqMetaType == EOF)
 						{
-							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", ftell(fSEQq));
+							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", (int) ftell(fSEQq));
 							goto quit_main_loop;
 						}
 						fputc(seqMetaType, fMIDI);
@@ -286,7 +286,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 							{
 								printf("\n");
 							}
-							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", ftell(fSEQq));
+							fprintf(stderr, "Error: Unexpected EOF at 0x%08X\n", (int) ftell(fSEQq));
 							goto quit_main_loop;
 						}
 						fputvb(seqMetaLength, fMIDI);
@@ -304,16 +304,16 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 						{
 							if (seqMetaLength > sizeof(s))
 							{
-								fprintf(stderr, "Error: Too long message at 0x%08X (the message may be valid, but this tool does not support it because of laziness.)\n", ftell(fSEQq));
+								fprintf(stderr, "Error: Too long message at 0x%08X (the message may be valid, but this tool does not support it because of laziness.)\n", (int) ftell(fSEQq));
 								goto quit_main_loop;
 							}
 
 							if (fread(s, seqMetaLength, 1, fSEQq) != 1) {
-								fprintf(stderr, "Error: File read error at 0x%08X\n", ftell(fSEQq));
+								fprintf(stderr, "Error: File read error at 0x%08X\n", (int) ftell(fSEQq));
 								goto quit_main_loop;
 							}
 							if (fwrite(s, seqMetaLength, 1, fMIDI) != 1) {
-								fprintf(stderr, "Error: File write error at 0x%08X\n", ftell(fMIDI));
+								fprintf(stderr, "Error: File write error at 0x%08X\n", (int) ftell(fMIDI));
 								goto quit_main_loop;
 							}
 							if (!quiet)
@@ -347,7 +347,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 					{
 						printf("- %06X: %06d %02X %02X\n", seqEventAddr, seqAbsTime, seqOpcode, seqByte);
 					}
-					fprintf(stderr, "Error: Unknown status 0x%02X at 0x%08X\n", seqOpcode, ftell(fMIDI));
+					fprintf(stderr, "Error: Unknown status 0x%02X at 0x%08X\n", seqOpcode, (int) ftell(fMIDI));
 					goto quit_main_loop;
 			}
 		}
@@ -357,11 +357,11 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 			if (seqEventLength > 0)
 			{
 				if (fread(s, seqEventLength, 1, fSEQq) != 1) {
-					fprintf(stderr, "Error: File read error at 0x%08X\n", ftell(fSEQq));
+					fprintf(stderr, "Error: File read error at 0x%08X\n", (int) ftell(fSEQq));
 					break;
 				}
 				if (fwrite(s, seqEventLength, 1, fMIDI) != 1) {
-					fprintf(stderr, "Error: File write error at 0x%08X\n", ftell(fMIDI));
+					fprintf(stderr, "Error: File write error at 0x%08X\n", (int) ftell(fMIDI));
 					break;
 				}
 				if (!quiet)
@@ -386,7 +386,7 @@ bool seqq2mid(FILE *fSEQq, FILE *fMIDI, bool quiet)
 quit_main_loop:
 
 	// set final track size
-	seqTrackSize = ftell(fMIDI) - 0x16;
+	seqTrackSize = (int) ftell(fMIDI) - 0x16;
 	fseek(fMIDI, 0x12, SEEK_SET);
 	fput4b(seqTrackSize, fMIDI);
